Create Main_Code
This code was utilized for the majority of the analysis presented in my paper. I have reviewed and refined it by removing redundant sections and elements that were intended for personal reference. Please note that there are instances where I alternated between subsetting the data and performing the analysis. I am providing this code as a reference framework. You may need to modify it to suit your specific requirements.